+++ lib/domain_extractor/result.rb
@@ -1,5 +1,7 @@
 # frozen_string_literal: true
 
+require_relative 'parsed_url'
+
 module DomainExtractor
   # Result encapsulates the final parsed attributes and exposes a hash interface.
   module Result
@@ -10,7 +12,7 @@ module DomainExtractor
     module_function
 
     def build(**attributes)
-      {
+      hash = {
         subdomain: normalize_subdomain(attributes[:subdomain]),
         root_domain: attributes[:root_domain],
         domain: attributes[:domain],
@@ -19,6 +21,8 @@ module DomainExtractor
         path: attributes[:path] || EMPTY_PATH,
         query_params: QueryParams.call(attributes[:query])
       }.freeze
+
+      ParsedURL.new(hash)
     end
 
     def normalize_subdomain(value)